GemIDService/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
fastapi==0.115.11
idna==3.10
PyJWT
pillow==11.1.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
python-multipart==0.0.20
requests==2.32.3
sniffio==1.3.1
starlette==0.46.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn
fastapi